+++ apps/server-go-htmx/.air.toml
@@ -5,7 +5,7 @@ tmp_dir = "tmp"
 testdata_dir = "testdata"
 
 [build]
-  cmd = "go build -o ./tmp/main ."     # Just plain old shell command. You could use `make` as well
+  cmd = "go build -o ./tmp/main ./cmd/api/main.go"         # Just plain old shell command. You could use `make` as well
   pre_cmd = []                         # Array of commands to run before each build
   post_cmd = []                        # Array of commands to run after ^C
   bin = "./tmp/main"                   # Binary file yields from `cmd`
@@ -16,7 +16,7 @@ testdata_dir = "testdata"
   exclude_file = []                    # Exclude files
   exclude_dir = ["assets", "tmp", "vendor", "testdata"]    # Ignore these filename extensions or directories.
   exclude_regex = ["_test.go"]         # Exclude specific regular expressions
-  include_ext = ["go", "tpl", "tmpl", "html"]              # Watch these filename extensions
+  include_ext = ["go", "html", "tpl", "tmpl"]              # Watch these filename extensions
   exclude_unchanged = true             # Exclude unchanged files
   follow_symlink = false               # Follow symlink for directories
   log = "build-errors.log"             # Log file in tmp_dir
